# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Pulse 109 Platform API
  version: 1.0.0
  description: >-
    Contract-first API for intake, operator assistance, incident management,
    analytics and regional system adapters. All timestamps use RFC 3339 with
    an explicit UTC offset. Human confirmation remains mandatory for routing,
    priority changes, duplicate merges and generated replies.
servers:
  - url: https://pulse109.example.gov.kz
    description: Production placeholder. Replace after infrastructure approval.
...
paths:
  /v1/appeals/preflight:
    post:
      tags: [Requests, Incidents]
      summary: Suggest possible duplicates before an appeal is submitted
      operationId: preflightAppeal
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RegionId'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreflightRequest'
      responses:
        '200':
          description: Advisory duplicate candidates; no appeal is stored or merged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreflightResult'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/Unprocessable'
...
  /v1/requests/{request_id}/similar:
    get:
      tags: [Requests]
      summary: Find similar resolved appeals and evidence-backed outcomes
      operationId: findSimilarRequests
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/CorrelationId'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
        - name: service_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Ranked candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SimilarRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/requests/{request_id}/duplicate-candidates:
    get:
      tags: [Incidents]
      summary: Find duplicate or incident membership candidates
      operationId: findDuplicateCandidates
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Ranked duplicate candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DuplicateCandidate'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by the approved government identity provider.
  parameters:
...
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
        maxLength: 128
    RegionId:
      name: X-Region-Id
      in: header
      required: true
      schema:
        type: string
        pattern: '^[A-Z0-9_-]{2,32}$'
    RequestId:
      name: request_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Forbidden:
      description: The verified actor lacks the required region or purpose scope
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Unprocessable:
      description: Business validation failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ServiceUnavailable:
      description: Optional AI or retrieval dependency is unavailable
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    SimilarRequest:
      type: object
      required: [request_id, score, evidence_type]
      properties:
        request_id:
          type: string
          format: uuid
        score:
          type: number
          minimum: 0
          maximum: 1
        evidence_type:
          type: string
          enum: [resolved_appeal, knowledge_article, service_definition]
        matched_fields:
          type: array
          items:
            type: string
        outcome_summary:
          type: string
        outcome_ref:
          type: string
    DuplicateCandidate:
      type: object
      required: [candidate_type, candidate_id, score, reasons, needs_human_confirmation]
      properties:
        candidate_type:
          type: string
          enum: [request, incident]
        candidate_id:
          type: string
          format: uuid
        score:
          type: number
          minimum: 0
          maximum: 1
        reasons:
          type: array
          items:
            type: string
        distance_m:
          type: [number, 'null']
        time_delta_minutes:
          type: [number, 'null']
        needs_human_confirmation:
          const: true
    PreflightRequest:
      type: object
      additionalProperties: false
      required: [region_id, service_id, topic_id, text]
      properties:
        region_id:
          type: string
          pattern: '^[A-Z0-9_-]{2,32}$'
        service_id:
          type: string
          minLength: 1
        topic_id:
          type: string
          minLength: 1
        text:
          type: string
          minLength: 1
          maxLength: 20000
        occurred_at:
          type: [string, 'null']
          format: date-time
        occurred_at_quality:
          type: string
          enum: [exact, source_tz_assumed, date_only, missing]
          default: missing
        latitude:
          type: [number, 'null']
          minimum: -90
          maximum: 90
        longitude:
          type: [number, 'null']
          minimum: -180
          maximum: 180
        limit:
          type: integer
          minimum: 1
          maximum: 10
          default: 5
    PreflightResult:
      type: object
      additionalProperties: false
      required:
        [candidates, evaluated_factors, needs_human_confirmation, automatic_merge, synthetic_only]
      properties:
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/DuplicateCandidate'
        evaluated_factors:
          type: array
          items:
            type: string
            enum: [category, distance, time, lexical, semantic]
        needs_human_confirmation:
          const: true
        automatic_merge:
          const: false
        synthetic_only:
          type: boolean
...
    Problem:
      type: object
      required: [type, title, status, trace_id]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        code:
          type: string
        trace_id:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string